Starting new prototype w/ skeleton project
