+++ Desktop/Desktop.iml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<module external.linked.project.path="$MODULE_DIR$" external.root.project.path="$MODULE_DIR$/.." external.system.id="GRADLE" external.system.module.group="tradesong" external.system.module.version="1.0" type="JAVA_MODULE" version="4">
+<module external.linked.project.id=":desktop" external.linked.project.path="$MODULE_DIR$" external.root.project.path="$MODULE_DIR$/.." external.system.id="GRADLE" external.system.module.group="tradesong" external.system.module.version="1.0" type="JAVA_MODULE" version="4">
   <component name="NewModuleRootManager" inherit-compiler-output="false">
     <output url="file://$MODULE_DIR$/build/classes/main" />
     <output-test url="file://$MODULE_DIR$/build/classes/test" />
@@ -15,25 +15,23 @@
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
     <orderEntry type="module" module-name="core" exported="" />
-    <orderEntry type="library" exported="" name="Gradle: gdx-backend-lwjgl-1.3.1" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: gdx-platform-1.3.1-natives-desktop" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: gdx-tools-1.3.1" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: gdx-controllers-desktop-1.3.1" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: gdx-controllers-platform-1.3.1-natives-desktop" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: gdx-1.3.1" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: lwjgl-platform-2.9.1-natives-linux" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: lwjgl-platform-2.9.1-natives-osx" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: lwjgl-platform-2.9.1-natives-windows" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: jutils-1.0.0" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: jinput-2.0.5" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: lwjgl-2.9.1" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: jinput-platform-2.0.5-natives-linux" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: jinput-platform-2.0.5-natives-osx" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: jinput-platform-2.0.5-natives-windows" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: lwjgl_util-2.9.1" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: jlayer-1.0.1-gdx" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: jorbis-0.0.17" level="project" />
-    <orderEntry type="library" exported="" name="Gradle: gdx-controllers-1.3.1" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: com.badlogicgames.gdx:gdx-backend-lwjgl:1.3.1" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: com.badlogicgames.gdx:gdx-platform:1.3.1:natives-desktop" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: com.badlogicgames.gdx:gdx-controllers-desktop:1.3.1" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: com.badlogicgames.gdx:gdx-controllers-platform:1.3.1:natives-desktop" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: com.badlogicgames.gdx:gdx:1.3.1" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: org.lwjgl.lwjgl:lwjgl-platform:2.9.1:natives-linux" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: org.lwjgl.lwjgl:lwjgl-platform:2.9.1:natives-osx" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: org.lwjgl.lwjgl:lwjgl-platform:2.9.1:natives-windows" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: net.java.jutils:jutils:1.0.0" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: net.java.jinput:jinput:2.0.5" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: org.lwjgl.lwjgl:lwjgl:2.9.1" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: net.java.jinput:jinput-platform:2.0.5:natives-linux" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: net.java.jinput:jinput-platform:2.0.5:natives-osx" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: net.java.jinput:jinput-platform:2.0.5:natives-windows" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: org.lwjgl.lwjgl:lwjgl_util:2.9.1" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: com.badlogicgames.jlayer:jlayer:1.0.1-gdx" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: org.jcraft:jorbis:0.0.17" level="project" />
+    <orderEntry type="library" exported="" name="Gradle: com.badlogicgames.gdx:gdx-controllers:1.3.1" level="project" />
   </component>
 </module>
-
