--- Desktop/src/com/minus7games/tradesong/desktop/DesktopLauncher.java
@@ -1,12 +1,0 @@
-package com.minus7games.tradesong.desktop;
-
-import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
-import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
-import com.minus7games.tradesong.Tradesong;
-
-public class DesktopLauncher {
-	public static void main (String[] arg) {
-		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
-		new LwjglApplication(new Tradesong(), config);
-	}
-}
