+++ Desktop/src/icbat/games/tradesong/desktop/DesktopLauncher.java
@@ -1,0 +1,12 @@
+package icbat.games.tradesong.desktop;
+
+import com.badlogic.gdx.backends.lwjgl.LwjglApplication;
+import com.badlogic.gdx.backends.lwjgl.LwjglApplicationConfiguration;
+import icbat.games.tradesong.TradesongGame;
+
+public class DesktopLauncher {
+	public static void main (String[] arg) {
+		LwjglApplicationConfiguration config = new LwjglApplicationConfiguration();
+		new LwjglApplication(new TradesongGame(), config);
+	}
+}
