+++ build.gradle
@@ -1,27 +1,25 @@
 buildscript {
     repositories {
         mavenCentral()
-        mavenLocal()
-        maven { url 'https://github.com/steffenschaefer/gwt-gradle-plugin/raw/maven-repo/' }
+        jcenter()
     }
     dependencies {
-        classpath 'de.richsource.gradle.plugins:gwt-gradle-plugin:0.4'
+        classpath 'de.richsource.gradle.plugins:gwt-gradle-plugin:0.5'
     }
 }
 
 allprojects {
     apply plugin: "eclipse"
     apply plugin: "idea"
-    sourceCompatibility = 1.8
+
     version = '1.0'
     ext {
-        appName = 'Tradesong'
+        appName = 'tradesong'
         gdxVersion = '1.3.1'
-        roboVMVersion = '0.0.13'
+        roboVMVersion = '0.0.14'
     }
 
     repositories {
-        mavenLocal()
         mavenCentral()
         maven { url "https://oss.sonatype.org/content/repositories/snapshots/" }
         maven { url "https://oss.sonatype.org/content/repositories/releases/" }
@@ -36,7 +34,6 @@ project(":desktop") {
         compile project(":core")
         compile "com.badlogicgames.gdx:gdx-backend-lwjgl:$gdxVersion"
         compile "com.badlogicgames.gdx:gdx-platform:$gdxVersion:natives-desktop"
-        compile "com.badlogicgames.gdx:gdx-tools:$gdxVersion"
         compile "com.badlogicgames.gdx:gdx-controllers-desktop:$gdxVersion"
         compile "com.badlogicgames.gdx:gdx-controllers-platform:$gdxVersion:natives-desktop"
     }
@@ -65,8 +62,6 @@ project(":core") {
     dependencies {
         compile "com.badlogicgames.gdx:gdx:$gdxVersion"
         compile "com.badlogicgames.gdx:gdx-controllers:$gdxVersion"
-		testCompile group: 'junit', name: 'junit', version: '4.+'
-        testCompile "org.mockito:mockito-all:1.9.5+"
     }
 }
