--- core/assets/items.json
@@ -1,32 +1,0 @@
-[
-    {
-        internalName: silkWyrms,
-        displayName: "Silk Wyrms",
-        description: "Giant glands make for good silk."
-    },
-    {
-        internalName: cloth_silk,
-        displayName: "Silken Cloth",
-        description: "Doesn't harm the Wyrms, but it's thread-count it a bit low..."
-    },
-    {
-        internalName: thread_silk,
-        displayName: "Silken Thread",
-        description: "Why did we have to make the cloth first?"
-    },
-    {
-        internalName: shirt_silk,
-        displayName: "Silken shirt",
-        description: "Peasants! Look upon me and despair."
-    },
-    {
-        internalName: apple,
-        displayName: "Apple",
-        description: "Just one will drive physicians crazy!"
-    },
-    {
-        internalName: orange,
-        displayName: "Orange",
-        description: "Definitely not an Apple."
-    }
-]
