--- core/assets/nodes.json
@@ -1,79 +1,0 @@
-[
-    {
-        internalName: orchard,
-        displayName: "Orchard",
-        possibleCraftSteps: [
-            {
-                displayName: "Grow Silk",
-                validInput: [
-                    {
-                        internalName: silkWyrms
-                    }
-                ],
-                outputs: [
-                    {
-                        internalName: cloth_silk
-                    }
-                ]
-            },
-            {
-                displayName: "Grow Apples",
-                validInput: [],
-                outputs: [
-                    {
-                        internalName: "apple"
-                    }
-                ]
-            },
-            {
-
-                displayName: "Grow Oranges",
-                validInput: [],
-                outputs: [
-                    {
-                        internalName: "orange"
-                    }
-                ]
-            }
-        ]
-
-    },
-    {
-        internalName: tailor,
-        displayName: "Tailor",
-        possibleCraftSteps: [
-            {
-                displayName: "Spin Silken Thread",
-                validInput: [
-                    {
-                        internalName: cloth_silk
-                    }
-                ],
-                outputs: [
-                    {
-                        internalName: thread_silk
-                    }
-
-                ]
-            },
-            {
-                displayName: "Sew Silken Shirt",
-                validInput: [
-                    {
-                        internalName: cloth_silk
-                    },
-                    {
-                        internalName: thread_silk
-                    }
-                ],
-                outputs: [
-                    {
-                        internalName: shirt_silk
-                    }
-
-                ]
-            }
-        ]
-    }
-]
-
