+++ core/build.gradle
@@ -1,6 +1,6 @@
 apply plugin: "java"
 
-sourceCompatibility = 1.8
+sourceCompatibility = 1.6
 [compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
 
 sourceSets.main.java.srcDirs = [ "src/" ]
