--- core/src/com/minus7games/tradesong/Displayable.java
@@ -1,8 +1,0 @@
-package com.minus7games.tradesong;
-
-import com.badlogic.gdx.scenes.scene2d.Actor;
-
-/***/
-public interface Displayable {
-    public Actor getActor();
-}
