+++ gradle.properties
@@ -1,3 +1,3 @@
 org.gradle.daemon=true
-org.gradle.jvmargs=-Xms128m -Xmx256m
+org.gradle.jvmargs=-Xms128m -Xmx512m
 org.gradle.configureondemand=true
