--- html/src/com/minus7games/tradesong/GdxDefinition.gwt.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit trunk//EN" "http://google-web-toolkit.googlecode.com/svn/trunk/distro-source/core/src/gwt-module.dtd">
-<module rename-to="html">
-	<inherits name='com.badlogic.gdx.backends.gdx_backends_gwt' />
-	<inherits name='com.badlogic.gdx.physics.box2d.box2d-gwt' />
-
-	<inherits name='MyGdxGame' />
-	<entry-point class='com.minus7games.tradesong.client.HtmlLauncher' />
-	
-	<set-configuration-property name="gdx.assetpath" value="../core/assets" />
-</module>
