--- html/src/com/minus7games/tradesong/client/HtmlLauncher.java
@@ -1,19 +1,0 @@
-package com.minus7games.tradesong.client;
-
-import com.badlogic.gdx.ApplicationListener;
-import com.badlogic.gdx.backends.gwt.GwtApplication;
-import com.badlogic.gdx.backends.gwt.GwtApplicationConfiguration;
-import com.minus7games.tradesong.Tradesong;
-
-public class HtmlLauncher extends GwtApplication {
-
-        @Override
-        public GwtApplicationConfiguration getConfig () {
-                return new GwtApplicationConfiguration(480, 320);
-        }
-
-        @Override
-        public ApplicationListener getApplicationListener () {
-                return new Tradesong();
-        }
-}
