+++ html/src/icbat/games/tradesong/GdxDefinitionSuperdev.gwt.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE module PUBLIC "-//Google Inc.//DTD Google Web Toolkit trunk//EN" "http://google-web-toolkit.googlecode.com/svn/trunk/distro-source/core/src/gwt-module.dtd">
+<module rename-to="html">
+	<inherits name='com.badlogic.gdx.backends.gdx_backends_gwt' />
+	<inherits name='com.badlogic.gdx.controllers.controllers-gwt' />
+
+    <inherits name='TradesongGame' />
+	<entry-point class='icbat.games.tradesong.client.HtmlLauncher' />
+	
+	<add-linker name="xsiframe"/>
+	<set-configuration-property name='xsiframe.failIfScriptTag' value='FALSE'/>
+	<set-configuration-property name="devModeRedirectEnabled" value="true"/>
+	
+	<set-configuration-property name="gdx.assetpath" value="../core/assets" />
+</module>
