+++ html/src/icbat/games/tradesong/client/HtmlLauncher.java
@@ -1,0 +1,19 @@
+package icbat.games.tradesong.client;
+
+import com.badlogic.gdx.ApplicationListener;
+import com.badlogic.gdx.backends.gwt.GwtApplication;
+import com.badlogic.gdx.backends.gwt.GwtApplicationConfiguration;
+import icbat.games.tradesong.TradesongGame;
+
+public class HtmlLauncher extends GwtApplication {
+
+        @Override
+        public GwtApplicationConfiguration getConfig () {
+                return new GwtApplicationConfiguration(480, 320);
+        }
+
+        @Override
+        public ApplicationListener getApplicationListener () {
+                return new TradesongGame();
+        }
+}
