+++ html/webapp/index.html
@@ -1,7 +1,7 @@
 <!doctype html>
 <html>
        <head>
-              <title>my-gdx-game</title>
+              <title>tradesong</title>
               <meta http-equiv="content-type" content="text/html; charset=UTF-8">
               <link href="styles.css" rel="stylesheet" type="text/css">
               <script src="soundmanager2-setup.js"></script>
